+++ e2e/e2e.go
@@ -3,7 +3,6 @@ package e2e
 import (
 	"context"
 	"errors"
-	"github.com/base-org/pessimism/internal/metrics"
 	"testing"
 	"time"
 
@@ -12,6 +11,7 @@ import (
 	"github.com/base-org/pessimism/internal/app"
 	"github.com/base-org/pessimism/internal/config"
 	"github.com/base-org/pessimism/internal/logging"
+	"github.com/base-org/pessimism/internal/metrics"
 	"github.com/base-org/pessimism/internal/state"
 	op_e2e "github.com/ethereum-optimism/optimism/op-e2e"
 	"github.com/ethereum/go-ethereum"
@@ -153,7 +153,7 @@ func DefaultTestConfig() *config.Config {
 			L1PollInterval: l1PollInterval,
 		},
 		MetricsConfig: &metrics.Config{
-			Enabled: true,
+			Enabled: false,
 			Host:    "localhost",
 			Port:    metPort,
 		},
