+++ internal/core/context.go
@@ -4,6 +4,7 @@ type CtxKey uint8
 
 const (
 	Logger CtxKey = iota
+	Metrics
 	State
 	L1Client
 	L2Client
